Img to texture function & refactoring
png leaks stopped by converting textures to images and then deleting them immediately, created new safe_mlx file, makefile now runs with all flags as all warnings are fixed
