+++ src/image_utils.c
@@ -6,7 +6,7 @@
 /*   By: tcampbel <tcampbel@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/17 15:00:00 by clundber          #+#    #+#             */
-/*   Updated: 2024/07/25 17:52:19 by tcampbel         ###   ########.fr       */
+/*   Updated: 2024/07/26 17:25:00 by tcampbel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	color_image(mlx_image_t *image, int color)
 {
-	int	x;
-	int	y;
+	uint32_t	x;
+	uint32_t	y;
 
 	x = 0;
 	while (x < image->width)
@@ -33,17 +33,4 @@ void	color_image(mlx_image_t *image, int color)
 int	make_color(int r, int g, int b, int a)
 {
 	return (r << 24 | g << 16 | b << 8 | a);
-}
-
-void	safe_pixel(mlx_image_t *img, uint32_t x, uint32_t y, uint32_t color)
-{
-	if (y < img->height && x < img->width)
-		mlx_put_pixel(img, x, y, color);
-}
-
-void	safe_image(t_data *data, uint32_t w, uint32_t h, mlx_image_t **img)
-{
-	*img = mlx_new_image(data->mlx, w, h);
-	if (!*img)
-		armageddon(data, "image mallocing failed");
 }
